% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGI0.R
\name{dGI0}
\alias{dGI0}
\title{Probability density function of the GI0 distribution}
\usage{
dGI0(x, p_alpha, p_gamma, p_Looks)
}
\arguments{
\item{p_alpha}{negative value that controls the roughness}

\item{p_gamma}{positive value that controls the scale}

\item{p_Looks}{value larger than 1 that controls the number of looks}
}
\description{
Probability density function of the GI0 distribution
}
\examples{
x <- seq(0,5,length.out=500)
plot(x,dGI0(x, p_alpha=-1, p_gamma=2, p_Looks=2), type='l')

}
\keyword{Density,}
\keyword{Distribution}
