% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGI0.R
\name{pGI0}
\alias{pGI0}
\title{Cumulative density function of the GI0 distribution}
\usage{
pGI0(x, p_alpha, p_gamma, p_Looks, log.p = FALSE)
}
\arguments{
\item{p_alpha}{negative value that controls the roughness}

\item{p_gamma}{positive value that controls the scale}

\item{p_Looks}{value larger than 1 that controls the number of looks}

\item{log.p}{optional logical parameter that switches to the logarithm}
}
\description{
Cumulative density function of the GI0 distribution
}
\examples{
x <- seq(0,5,length.out=500)
plot(x,pGI0(x, p_alpha=-1, p_gamma=2, p_Looks=2), type='l')
}
\keyword{Distribution}
