% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGI0.R
\name{qGI0}
\alias{qGI0}
\title{Quantile  function of the GI0 distribution}
\usage{
qGI0(p, p_alpha, p_gamma, p_Looks, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p_alpha}{negative value that controls the roughness}

\item{p_gamma}{positive value that controls the scale}

\item{p_Looks}{value larger than 1 that controls the number of looks}

\item{lower.tail}{optional logical parameter that computes only the lower tail}

\item{log.p}{optional logical parameter that switches to the logarithm}
}
\description{
Quantile  function of the GI0 distribution
}
\examples{
p <- seq(0,1,length.out=100)
plot(p,qGI0(p, p_alpha=-1, p_gamma=2, p_Looks=2), type='l')
}
\keyword{Distribution,}
\keyword{Quantile}
