% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGI0.R
\name{rGI0}
\alias{rGI0}
\title{Pseudorandom random number generator for the GI0 distribution}
\usage{
rGI0(n, p_alpha, p_gamma, p_Looks, from.F = FALSE)
}
\arguments{
\item{n}{the sample size}

\item{p_alpha}{negative value that controls the roughness}

\item{p_gamma}{positive value that controls the scale}

\item{p_Looks}{value larger than 1 that controls the number of looks}

\item{from.F}{optional logical parameter that sets the type of generator}
}
\description{
Pseudorandom random number generator for the GI0 distribution
}
\examples{
set.seed(123)
rGI0(10, p_alpha=-1, p_gamma=2, p_Looks=2)
}
