% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGI0.R
\name{dGI0}
\alias{dGI0}
\title{"PDF"}
\usage{
dGI0(x, p_alpha, p_gamma, p_Looks)
}
\arguments{
\item{p_alpha}{Negative value that controls the roughness}

\item{p_gamma}{Positive value that controls the scale}

\item{p_Looks}{Positive value that controls the number of looks}
}
\description{
"PDF"
}
\examples{
x<- seq(0,5,length.ou=500)
plot (x, dGI0(x, -2, 4, 5),type="l")

}
