% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGI0.R
\name{pGI0}
\alias{pGI0}
\title{CDF}
\usage{
pGI0(x, p_alpha, p_gamma, p_Looks, log.p = FALSE)
}
\arguments{
\item{p_alpha}{Negative value that controls the roughness}

\item{p_gamma}{Positive value that controls the scale}

\item{p_Looks}{Positive value that controls the number of looks}
}
\description{
CDF
}
