% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGI0.R
\name{qGI0}
\alias{qGI0}
\title{Quantile function}
\usage{
qGI0(p, p_alpha, p_gamma, p_Looks, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p_alpha}{Negative value that controls the roughness}

\item{p_gamma}{Positive value that controls the scale}

\item{p_Looks}{Positive value that controls the number of looks}

\item{lower.tatil}{Optional parameter that computes only the lower tail}

\item{log}{optional parameter that switches logarithm}
}
\description{
Quantile function
}
