% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGI0.R
\name{rGI0}
\alias{rGI0}
\title{Random number generator}
\usage{
rGI0(n, p_alpha, p_gamma, p_Looks, from.F = FALSE)
}
\arguments{
\item{p_alpha}{Negative value that controls the roughness}

\item{p_gamma}{Positive value that controls the scale}

\item{p_Looks}{Positive value that controls the number of looks}
}
\description{
Random number generator
}
